+++ src/grandorgue/combinations/model/GOCombinationElement.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2023 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -10,10 +10,9 @@
 
 class GOCombinationElement {
 public:
-  virtual ~GOCombinationElement() {}
-
   virtual bool GetCombinationState() const = 0;
   virtual void SetCombination(bool on) = 0;
+  virtual bool IsControlledByUser() const = 0;
 };
 
 #endif